Avoiding a few more warnings with GCC -Wconversion
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@1212 2e953b5c-c64d-0410-be54-f773e93e544c

+++ gouct/GoUctPlayer.h
@@ -64,7 +64,7 @@ public:
 
         SgStatisticsExt<float,std::size_t> m_reuse;
 
-        SgStatisticsExt<float,std::size_t> m_gamesPerSecond;
+        SgStatisticsExt<double,std::size_t> m_gamesPerSecond;
 
         Statistics();
